+++ Sources/PayPalWebPayments/PayPalWebCheckoutClientError.swift
@@ -22,7 +22,13 @@ enum PayPalWebCheckoutClientError {
         case malformedResultError
 
         /// 4. Vault result did not return a token id
-        case paypalVaultResponseError
+        case payPalVaultResponseError
+
+        /// 5. Websession is cancelled by the user
+        case payPalCancellationError
+
+        /// 6. Websession is cancelled by the user
+        case payPalVaultCancellationError
     }
 
     static let webSessionError: (Error) -> CoreSDKError = { error in
@@ -46,8 +52,20 @@ enum PayPalWebCheckoutClientError {
     )
 
     static let payPalVaultResponseError = CoreSDKError(
-        code: Code.paypalVaultResponseError.rawValue,
+        code: Code.payPalVaultResponseError.rawValue,
         domain: domain,
-        errorDescription: "Error parsing paypal vault response"
+        errorDescription: "Error parsing PayPal vault response"
+    )
+
+    static let payPalCancellationError = CoreSDKError(
+        code: Code.payPalCancellationError.rawValue,
+        domain: domain,
+        errorDescription: "PayPal checkout has been cancelled by the user"
+    )
+
+    static let payPalVaultCancellationError = CoreSDKError(
+        code: Code.payPalVaultCancellationError.rawValue,
+        domain: domain,
+        errorDescription: "PayPal vault has been cancelled by the user"
     )
 }